+++ wandb/apis/internal.py
@@ -170,9 +170,6 @@ class Api:
     def upload_multipart_file_chunk_retry(self, *args, **kwargs):
         return self.api.upload_multipart_file_chunk_retry(*args, **kwargs)
 
-    async def upload_file_retry_async(self, *args, **kwargs):
-        return await self.api.upload_file_retry_async(*args, **kwargs)
-
     def get_run_info(self, *args, **kwargs):
         return self.api.get_run_info(*args, **kwargs)