+++ CHANGELOG.md
@@ -33,6 +33,10 @@ Please add to the relevant subsections under Unreleased below on every PR where 
 * When using `wandb-core` need to specify a required flag (`wandb.require("core")`) to enable it, before it was picked up automatically by @kptkin in  https://github.com/wandb/wandb/pull/7228
 * Use ETags instead of MD5 hashes for GCS reference artifacts by @moredatarequired in https://github.com/wandb/wandb/pull/7337
 
+### Removed
+
+* Removed the `[async]` extra and the `_async_upload_concurrency_limit` setting by @moredatarequired in https://github.com/wandb/wandb/pull/7416
+
 ## [0.16.6] - 2024-04-03
 
 ### Added